TST-289 Reimplemented tsAddToCart
- to account for the add to cart button for
each of the products in the search result
- Github issue #419
