+++ templates/components/products/card.html
@@ -58,11 +58,19 @@
                         {{#if add_to_cart_url }}
                             {{!-- TST-244 Stop add to cart button from redirecting to cart page
                               - Added add-to-cart-btn class
-                              - Added a hidden input with its product_id value to be used
-                                in assets/js/theme/common/ts-add-to-cart.js
+
+                              TST-289
+                              - added onclick and data-product-id attribute
                             --}}
-                            <a href="{{add_to_cart_url}}" data-event-type="product-click" class="button-primary button--small add-to-cart-btn">{{lang 'products.add_to_cart'}}</a>
-                            <input type="hidden" name="product_id" value="{{product_id}}">
+                            <a
+                              href="{{add_to_cart_url}}"
+                              data-event-type="product-click"
+                              class="button-primary button--small add-to-cart-btn"
+                              onclick="tsAddToCart(this, event)"
+                              data-product-id="{{id}}"
+                            >
+                              {{lang 'products.add_to_cart'}}
+                            </a>
                         {{/if}}
                         {{#if out_of_stock_message }}
                             <a href="{{url}}" data-event-type="product-click" class="out-of-stock-button card-figcaption-button" data-product-id="{{id}}">
